requirements.txt
pymilvus
openai
requests
tqdm
python-dotenv==1.0.1
openai==1.51.2
langchain==0.2.16
streamlit
tiktoken==0.7.0
prometheus_client==0.21.0